Fix stall when pynini is imported before this library (#33)
* Fix stall when pynini is imported before this library

Previously, doing this:

import pynini
import riva.asrlib.decoder

would end in a stall. This was because the FlagRegister template class
had a static variable Mutex, that ended up being shared by default in
the global symbol namespace. Because the openfst versions are
different between the two libraries (I think pynini's openfst 1.8.2
uses absl::Mutex), the mutex layouts are different, probably causing
the stall.

We don't want the symbols to be shared in this case, so we disable
unique symbols via "-fno-gnu-unique".

* Increase error margin.

Helps with flaky CI failures.